Use new features introduced in branch_distributed_grids even in mainline.
git-svn-id: https://svn.dealii.org/trunk@20481 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/lac/source/petsc_matrix_base.cc
@@ -620,7 +620,7 @@ namespace PETScWrappers
 	std::cout << "*** PETSC-Matrix: num-allocs = "
 		  << info.mallocs << " ***" << std::endl;
 
-      return sizeof(*this) + info.memory;
+      return sizeof(*this) + static_cast<unsigned int>(info.memory);
   }
 
 }
